feat(task-006): slim smart_agent.py facade and extract registry scanner
- Extract ~600 lines of static registry scanning code to
  ovagent/registry_scanner.py (RegistryScanner class)
- Remove compatibility module-wiring block (handled by test fixture)
- Remove duplicate one-liner delegates that were not protocol-required
- Remove inline _search_functions/_get_function_details (already in ToolRuntime)
- Remove redundant _CODE_QUALITY_RULES alias, _initialize_skill_registry instance method
- Reduce smart_agent.py from 2336 to ~1345 lines (~42% reduction)
- Add lazy _scanner property for __new__-constructed agents in tests
- Add comprehensive test_ovagent_registry_scanner.py (22 tests)
- All 63 existing tests pass, 2 skipped (harness-only)

Acceptance criteria: smart_agent.py is a focused facade/composer;
imports and module boundaries are coherent; public APIs and regression
checks continue to pass.
